+++ .pre-commit-config.yaml
@@ -11,7 +11,6 @@ repos:
         args: [--remove]
       - id: check-added-large-files
         files: '^.*\.py$'
-      - id: imagetag-branch-constraint
   - repo: https://github.com/adrienverge/yamllint.git
     rev: v1.29.0
     hooks:
