+++ python_hooks/image_tag_branch_constraint.py
