+++ tests/python/data/image_tag_branch_sample.py
